Seperate files from `__init__.py` (#13)
* Seperate files from

* fix version again

* include import for easier user access

* Update __init__.py

add missing double quotes

* Make more formatting changes and update sanity-test.py
