+++ cohere/generation.py
@@ -1,0 +1,6 @@
+class Generation:
+    def __init__(self, text: str) -> None:
+        self.text = text
+    
+    def __str__(self) -> str:
+        return self.text
